+++ src/windows/common.rs
@@ -115,15 +115,15 @@ pub unsafe fn convert(param: WPARAM, lpdata: LPARAM) -> (Option<EventType>, u16)
             Ok(WM_MOUSEWHEEL) => {
                 let delta = get_delta(lpdata) as c_short;
                 Some(EventType::Wheel {
-                    delta_x: 0,
-                    delta_y: (delta / WHEEL_DELTA) as i64,
+                    delta_x: 0.0,
+                    delta_y: (delta as f64) / (WHEEL_DELTA as f64),
                 })
             }
             Ok(WM_MOUSEHWHEEL) => {
                 let delta = get_delta(lpdata) as c_short;
                 Some(EventType::Wheel {
-                    delta_x: (delta / WHEEL_DELTA) as i64,
-                    delta_y: 0,
+                    delta_x: (delta as f64) / (WHEEL_DELTA as f64),
+                    delta_y: 0.0,
                 })
             }
             _ => None,
